backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async Support
httpx>=0.26.0
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.26.0

# ML / Embeddings
sentence-transformers>=2.3.0
torch>=2.1.0

# LLM
google-generativeai>=0.3.2

# Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Caching
cachetools>=5.3.2

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
